refactor initializing igdb client in a separate file
